+++ pybop/costs/_likelihoods.py
@@ -8,38 +8,9 @@ class BaseLikelihood(BaseCost):
     Base class for likelihoods
     """
 
-    def __init__(self, problem, sigma=None):
-        super(BaseLikelihood, self).__init__(problem, sigma)
+    def __init__(self, problem):
+        super(BaseLikelihood, self).__init__(problem)
         self.n_time_data = problem.n_time_data
-
-    def set_sigma(self, sigma):
-        """
-        Setter for sigma parameter
-        """
-
-        if not isinstance(sigma, np.ndarray):
-            sigma = np.array(sigma)
-
-        if not np.issubdtype(sigma.dtype, np.number):
-            raise ValueError("Sigma must contain only numeric values")
-
-        if np.any(sigma <= 0):
-            raise ValueError("Sigma must not be negative")
-        else:
-            self.sigma0 = sigma
-
-    def get_sigma(self):
-        """
-        Getter for sigma parameter
-        """
-        return self.sigma0
-
-    @property
-    def n_parameters(self):
-        """
-        Returns the number of parameters
-        """
-        return self._n_parameters
 
 
 class GaussianLogLikelihoodKnownSigma(BaseLikelihood):
@@ -48,18 +19,49 @@ class GaussianLogLikelihoodKnownSigma(BaseLikelihood):
     which assumes that the data follows a Gaussian distribution and computes
     the log-likelihood of observed data under this assumption.
 
-    Attributes:
-        _logpi (float): Precomputed offset value for the log-likelihood function.
+    Parameters
+    ----------
+    sigma : scalar or array
+        Initial standard deviation around ``x0``. Either a scalar value (one
+        standard deviation for all coordinates) or an array with one entry
+        per dimension. Not all methods will use this information.
     """
 
     def __init__(self, problem, sigma):
-        super(GaussianLogLikelihoodKnownSigma, self).__init__(problem, sigma)
-        if sigma is not None:
-            self.set_sigma(sigma)
-        self._offset = -0.5 * self.n_time_data * np.log(2 * np.pi / self.sigma0)
-        self._multip = -1 / (2.0 * self.sigma0**2)
-        self.sigma2 = self.sigma0**-2
-        self._dl = np.ones(self._n_parameters)
+        super(GaussianLogLikelihoodKnownSigma, self).__init__(problem)
+        self.sigma = None
+        self.set_sigma(sigma)
+        self._offset = -0.5 * self.n_time_data * np.log(2 * np.pi / self.sigma)
+        self._multip = -1 / (2.0 * self.sigma**2)
+        self.sigma2 = self.sigma**-2
+        self._dl = np.ones(self.n_parameters)
+
+    def set_sigma(self, sigma):
+        """
+        Setter for sigma parameter
+        """
+        if sigma is None:
+            raise ValueError(
+                "The GaussianLogLikelihoodKnownSigma cost requires sigma to be "
+                + "either a scalar value or an array with one entry per dimension."
+            )
+
+        if not isinstance(sigma, np.ndarray):
+            sigma = np.array(sigma)
+
+        if not np.issubdtype(sigma.dtype, np.number):
+            raise ValueError("Sigma must contain only numeric values")
+
+        if np.any(sigma <= 0):
+            raise ValueError("Sigma must be positive")
+        else:
+            self.sigma = sigma
+
+    def get_sigma(self):
+        """
+        Getter for sigma parameter
+        """
+        return self.sigma
 
     def _evaluate(self, x, grad=None):
         """
@@ -112,14 +114,16 @@ class GaussianLogLikelihood(BaseLikelihood):
     data follows a Gaussian distribution and computes the log-likelihood of
     observed data under this assumption.
 
-    Attributes:
-        _logpi (float): Precomputed offset value for the log-likelihood function.
+    Attributes
+    ----------
+    _logpi : float
+        Precomputed offset value for the log-likelihood function.
     """
 
     def __init__(self, problem):
         super(GaussianLogLikelihood, self).__init__(problem)
         self._logpi = -0.5 * self.n_time_data * np.log(2 * np.pi)
-        self._dl = np.ones(self._n_parameters + self.n_outputs)
+        self._dl = np.ones(self.n_parameters + self.n_outputs)
 
     def _evaluate(self, x, grad=None):
         """
@@ -195,25 +199,19 @@ class LogPosterior(BaseCost):
     Inherits all parameters and attributes from ``BaseCost``.
     """
 
-    def __init__(self, log_likelihood, log_prior=None, sigma=None):
-        super(LogPosterior, self).__init__(problem=log_likelihood.problem, sigma=sigma)
+    def __init__(self, log_likelihood, log_prior=None):
+        super(LogPosterior, self).__init__(problem=log_likelihood.problem)
 
         # Store the likelihood and prior
         self._log_likelihood = log_likelihood
         self._prior = log_prior
         if self._prior is None:
             try:
-                self._prior = [
-                    param.prior
-                    for i, param in enumerate(
-                        log_likelihood.problem.parameters
-                    )  # Update for parameters class
-                ]
+                self._prior = log_likelihood.problem.parameters.priors()
             except Exception as e:
                 raise ValueError(
                     f"An error occurred when constructing the Prior class: {e}"
                 )
-
         try:  # This is a patch, the n_parameters val needs to be updated across the codebase
             self._n_parameters = self._prior.n_parameters
         except AttributeError: