+++ tests/unit/test_posterior.py
@@ -14,14 +14,17 @@ class TestLogPosterior:
         return pybop.lithium_ion.SPM()
 
     @pytest.fixture
-    def parameters(self):
-        return [
-            pybop.Parameter(
-                "Negative electrode active material volume fraction",
-                prior=pybop.Gaussian(0.5, 0.01),
-                bounds=[0.375, 0.625],
-            ),
-        ]
+    def ground_truth(self):
+        return 0.52
+
+    @pytest.fixture
+    def parameters(self, ground_truth):
+        return pybop.Parameter(
+            "Negative electrode active material volume fraction",
+            prior=pybop.Gaussian(0.5, 0.01),
+            bounds=[0.375, 0.625],
+            initial_value=ground_truth,
+        )
 
     @pytest.fixture
     def experiment(self):
@@ -32,15 +35,11 @@ class TestLogPosterior:
         )
 
     @pytest.fixture
-    def x0(self):
-        return np.array([0.52])
-
-    @pytest.fixture
-    def dataset(self, model, experiment, x0):
+    def dataset(self, model, experiment, ground_truth):
         model.parameter_set = model.pybamm_model.default_parameter_values
         model.parameter_set.update(
             {
-                "Negative electrode active material volume fraction": x0[0],
+                "Negative electrode active material volume fraction": ground_truth,
             }
         )
         solution = model.predict(experiment=experiment)
@@ -53,11 +52,8 @@ class TestLogPosterior:
         )
 
     @pytest.fixture
-    def one_signal_problem(self, model, parameters, dataset, x0):
-        signal = ["Voltage [V]"]
-        return pybop.FittingProblem(
-            model, parameters, dataset, signal=signal, x0=x0, init_soc=1.0
-        )
+    def one_signal_problem(self, model, parameters, dataset):
+        return pybop.FittingProblem(model, parameters, dataset, init_soc=1.0)
 
     @pytest.fixture
     def likelihood(self, one_signal_problem):
@@ -72,15 +68,8 @@ class TestLogPosterior:
         # Test log posterior construction
         posterior = pybop.LogPosterior(likelihood, prior)
 
-        assert posterior.sigma0 is not None
         assert posterior._log_likelihood == likelihood
         assert posterior._prior == prior
-
-        # Test log posterior construction without sigma
-        likelihood.sigma0 = None
-        likelihood.problem.sigma0 = None
-        posterior = pybop.LogPosterior(likelihood, prior, sigma=None)
-        assert posterior.sigma0 is not None
 
         # Test log posterior construction without parameters
         likelihood.problem.parameters = None
@@ -95,9 +84,8 @@ class TestLogPosterior:
         posterior = pybop.LogPosterior(likelihood, None)
 
         assert posterior._prior is not None
-        assert (
-            posterior._prior[0] == posterior._log_likelihood.problem.parameters[0].prior
-        )
+        for i, p in enumerate(posterior._prior):
+            assert p == posterior._log_likelihood.problem.parameters.priors()[i]
 
     @pytest.fixture
     def posterior(self, likelihood, prior):